+++ pkg/fields/fields_test.go
@@ -1,7 +1,7 @@
 // Copyright 2019-2022 Graham Clark. All rights reserved.  Use of this source code is governed by the MIT license
 // that can be found in the LICENSE file.
 
-package termshark
+package fields
 
 import (
 	"testing"
@@ -13,7 +13,7 @@ import (
 
 func TestFields1(t *testing.T) {
 
-	fields := NewFields()
+	fields := New()
 	err := fields.InitNoCache()
 	assert.NoError(t, err)