+++ ui/searchalg.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/gcla/gowid"
 	"github.com/gcla/termshark/v2"
-	"github.com/gcla/termshark/v2/pcap"
+	"github.com/gcla/termshark/v2/pkg/pcap"
 	"github.com/gcla/termshark/v2/widgets/search"
 )